feat(): Infer return type of ConfigService.get method
The ConfigService class has a generic template which allows to infer the
return type when getting a property.

BREAKING CHANGE:

The generic type of the get method is not the return type anymore.
This generic type is now the key in config. Most of the time this
generic parameter can be omit.