test(): unexpected stringification of variables assigned to process.env
After validation, environment variables are assigned back to
process.env. This leads to stringification of the values. This
stringification is deprecated and might result in a runtime error in
the future. See https://nodejs.org/api/process.html#processenv.

A special case is when an environment variable has the value
`undefined`. This causes `ConfigService.get` to fall back on
process.env thereby returning the stringified form `'undefined'`.